refactor: extract personalization routes
Refs #826
