+++ backend/news_dashboard/personalization/__init__.py
@@ -1,0 +1,1 @@
+"""Personalization feature module."""
